backend/requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2
django-cors-headers==4.7.0
gunicorn==23.0.0
idna==3.11
packaging==25.0
pillow==12.0.0
python-dotenv==1.2.1
requests==2.32.4
sqlparse==0.5.3
urllib3==2.5.0
psycopg[binary]
django-modeltranslation
